% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_create_field_level_encryption_config}
\alias{cloudfront_create_field_level_encryption_config}
\title{Create a new field-level encryption configuration}
\usage{
cloudfront_create_field_level_encryption_config(
  FieldLevelEncryptionConfig)
}
\arguments{
\item{FieldLevelEncryptionConfig}{[required] The request to create a new field-level encryption configuration.}
}
\description{
Create a new field-level encryption configuration.
}
\section{Request syntax}{
\preformatted{cloudfront$create_field_level_encryption_config(
  FieldLevelEncryptionConfig = list(
    CallerReference = "string",
    Comment = "string",
    QueryArgProfileConfig = list(
      ForwardWhenQueryArgProfileIsUnknown = TRUE|FALSE,
      QueryArgProfiles = list(
        Quantity = 123,
        Items = list(
          list(
            QueryArg = "string",
            ProfileId = "string"
          )
        )
      )
    ),
    ContentTypeProfileConfig = list(
      ForwardWhenContentTypeIsUnknown = TRUE|FALSE,
      ContentTypeProfiles = list(
        Quantity = 123,
        Items = list(
          list(
            Format = "URLEncoded",
            ProfileId = "string",
            ContentType = "string"
          )
        )
      )
    )
  )
)
}
}

\keyword{internal}
